feat(Footer): add data
