+++ src/components/layouts/Footer.tsx
@@ -1,6 +1,29 @@
 import React from 'react'
-import { Box } from 'danni-s-design-system'
+
+import { HoverableText, Text, Link } from 'danni-s-design-system'
+
+import { CONTACTS } from 'constants/contacts'
+
+const TTO = 2021
 
 export const Footer: React.FC = () => {
-  return <Box>Hello world, {new Date().getFullYear()}</Box>
+  const currentYear = new Date().getFullYear()
+
+  return (
+    <Text mr="xxl" sx={{ direction: 'rtl' }}>
+      <Text color="accentLightest" ml="s" inlineBlock>
+        2021 {currentYear !== TTO && `– ${currentYear}`}
+      </Text>
+      <Link
+        color="accentLightest"
+        href={CONTACTS.GITHUB.link}
+        sx={{ textDecoration: 'none' }}
+        inlineBlock
+      >
+        <HoverableText activeColour="complementaryDark">
+          © {CONTACTS.GITHUB.value}
+        </HoverableText>
+      </Link>
+    </Text>
+  )
 }
