+++ types/index.d.ts
@@ -1,3 +1,4 @@
-export * from './Page'
-export * from './Locale'
 export * from './Contact'
+export * from './Locale'
+export * from './MeFact'
+export * from './Page'
